Rename Input_config, Input_task
