+++ card_creator_config.py
@@ -1,7 +1,11 @@
 from kaiten.user import User
 import json
 
-class Input_config:
+# TODO: Сделать json методом, чтобы он находу собирал объект,
+# чтобы не обращаться к object.config['member'] вместо object.member 
+#
+# TODO: Добавить список тегов в конфиг, чтобы он не был захардкожен
+class Card_creator_config:
     """!Конфигурация для создания карточек: доска, колонка, роль и тип карточки"""
 
     def __init__(self, config_name: str, owner: User, size: int = None):
@@ -9,12 +13,9 @@ class Input_config:
         self.config['owner_id'] = owner.id
         self.config['owner_email'] = owner.email
         if size is not None:
-            self.config.json['size_text'] = f"{size} ч"
-
+            self.config['size_text'] = f"{size} ч"
         for key in self.config:
             setattr(self, key, self.config[key])
-        self.owner_id = owner.id
-        self.owner_email = owner.email
 
     @property
     def json(self):
